+++ src/app/shared/shared.module.ts
@@ -1,18 +1,21 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import {MouseOverHighlightDirective} from '../directive/mouse-over-highlight.directive';
+import { OrderByPipe } from './order-by.pipe';
 
 @NgModule({
   imports: [
     CommonModule
   ],
   exports: [
-    MouseOverHighlightDirective
+    MouseOverHighlightDirective,
+    OrderByPipe
   ],
   providers: [
   ],
   declarations: [
-    MouseOverHighlightDirective
+    MouseOverHighlightDirective,
+    OrderByPipe
   ]
 })
 export class SharedModule { }